feat(310): refactor category config list into own component
